micromegas_tracing

Module dispatch

Where events are recorded and eventually sent to a sink

Re-exports

pub use crate::errors::Error;

pub use crate::errors::Result;

Structs

DispatchCell

Functions

cpu_tracing_enabled

float_metric

flush_log_buffer

flush_metrics_buffer

flush_thread_buffer

for_each_thread_stream

force_uninit⚠

Safety

get_sink

init_event_dispatch

init_thread_stream

int_metric

log

log_enabled

log_interop

log_tagged

make_process_info

on_begin_async_named_scope

on_begin_async_scope

on_begin_named_scope

on_begin_scope

on_end_async_named_scope

on_end_async_scope

on_end_named_scope

on_end_scope

process_id

shutdown_dispatch

tagged_float_metric

tagged_integer_metric

unregister_thread_stream

Unregisters the current thread’s stream from the global dispatch to prevent dangling pointers when the thread is destroyed.